+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 dd mm yyyy
 
 - New: CSV export for charts.
+- New: The warning about disabled groups now distinguishes between disabled and soon to be disabled.
 - Fix: JavaScript error when adding a new EVE login.
 - Text changes/improvements.